backend/requirements.txt
# Core dependencies
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
gunicorn>=21.2.0

# Database
sqlalchemy>=2.0.25
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Face recognition
face_recognition>=1.3.0
numpy>=1.26.3
pillow>=10.1.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.2
httpx>=0.25.2

# Utilities
python-dateutil>=2.8.2
tenacity>=8.2.3
starlette>=0.27.0
psutil>=5.9.0